+++ extensions/how-long-to-beat/src/constants.ts
@@ -1,7 +1,7 @@
 export const HLTB_BASE_URL = "https://howlongtobeat.com/";
 export const HLTB_DETAIL_URL = `${HLTB_BASE_URL}game/`;
 export const HLTB_IMAGE_URL = `${HLTB_BASE_URL}games/`;
-export const HLTB_API_SEARCH_ENDPOINT = "/api/search/";
-export const HLTB_API_SEARCH_INIT_ENDPOINT = "/api/search/init";
+export const HLTB_API_SEARCH_ENDPOINT = "/api/finder/";
+export const HLTB_API_SEARCH_INIT_ENDPOINT = "/api/finder/init";
 
 export const TOKEN_CACHE_DURATION_MS = 5 * 60 * 1000;